Finalize the seeding script
