+++ server/config/init.js
@@ -1,0 +1,7 @@
+module.exports = {
+  startId: 1,
+  endId: 100,
+  localDir: 'images',
+  remoteDir: 'medium',
+  remotePath: 'https://product-detail-carousel-images.s3-us-west-1.amazonaws.com/',
+};
